[MOL-21341][XZY] Refactor NDI OIDC V2 types and update unit tests
